- Improved 2D function evaluation. - Improvements to worksheet generation from plot curves.
git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@1497 5a6a7de5-fb12-0410-b871-c33778c25c60

+++ qtiplot/src/plot2D/LineProfileTool.h
@@ -77,7 +77,7 @@ class LineProfileTool : public QWidget, public PlotToolInterface
 		void statusText(const QString&);
 
 	protected:
-		int averageImagePixel(const QImage &image, int px, int py, bool moreHorizontal);
+		double averageImagePixel(const QImage &image, int px, int py, bool moreHorizontal);
 		/*!\brief Draw line during operation.
 		 */
 		virtual void paintEvent(QPaintEvent *e);
